procedures CRUD has been created

+++ src/main/java/dev/downloadablefox/tabbies/webserver/entities/Medicine.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 public class Medicine {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     
     @Column(nullable = false)
     private String name;
@@ -26,7 +26,7 @@ public class Medicine {
     @Column(nullable = false)
     private Integer sold;
 
-    public Medicine(Integer id, String name, Double buyPrice, Double sellPrice, Integer stock, Integer sold) {
+    public Medicine(Long id, String name, Double buyPrice, Double sellPrice, Integer stock, Integer sold) {
         this.id = id;
         this.name = name;
         this.buyPrice = buyPrice;
@@ -45,11 +45,11 @@ public class Medicine {
 
     public Medicine() {}
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
